requirements.txt
click==8.2.1
commonmark==0.9.1
Deprecated==1.2.18
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Limiter==2.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
MarkupSafe==3.0.2
packaging==25.0
Pygments==2.19.1
PyJWT==2.3.0
rich==12.6.0
six==1.17.0
typing_extensions==4.13.2
waitress==2.1.1
Werkzeug==2.0.3
wrapt==1.17.2
